First round attempt to add one additional pasture blanket to the SWEM mod.
Notes about this attempt:
Was able to launch game and load in prior to making new TrePastureBlanketItem class. The launch with the new SWEM blanket was able to load the item and apply it to the horse and render it... just missing textures.

TrePastureBlanketItem is meant to change the location of the resource looked at for the PastureBlanketItem from SWEM. This is throwing an error code (commented error code)

Will do more research and learning and reattempt later as it is 2:30am