Added all halter textures for Halter update. -- Rainbow Halter (On Horse Render) by Moon Flower
Each Halter has 3 variants, a render on the horse, a render on the bridle rack, and a GUI icon.

Halters 1-9 are normal halters, and halters 10-12 are designed to look like flymasks. Flymask halters will not display like flymasks on the bridle rack, but rather like a normal halter due to how the HalterItem.class is written for SWEM.